build: refactoring sphinx build logic and changing approach to fabric use

+++ bin/makecloth/sphinx.py
@@ -15,22 +15,6 @@ from makecloth import MakefileCloth
 m = MakefileCloth()
 
 def make_all_sphinx(sphinx):
-    m.section_break('sphinx related variables', block='header')
-
-    m.comment('variables related to paper sizing for the latex output', block='vars')
-    m.newline()
-    m.var(variable='PAPER',
-          value='letter',
-          block='vars')
-    m.var(variable='PAPEROPT_a4',
-          value='-D latex_paper_size=a4',
-          block='vars')
-    m.var(variable='PAPEROPT_letter',
-          value='-D latex_paper_size=letter',
-          block='vars')
-
-    m.comment('general sphinx variables', block='vars')
-
     m.section_break('sphinx prerequisites')
     m.newline()
     m.target('sphinx-prerequisites', 'setup composites generate-source composite-pages.yaml', block='prereq')
@@ -72,44 +56,23 @@ def sphinx_builder(target):
     m.append_var('sphinx-targets', target)
     m.target(target, 'sphinx-prerequisites', block=b)
 
-    build_kickoff(target, block=b)
-
-    if target.startswith('html') or target.startswith('dirhtml'):
-        tag = 'website'
-    else:
-        tag = 'print'
-
-    sphinx_opts = '-q -d $(branch-output)/doctrees-$@ $(PAPEROPT_$(PAPER)) -c ./ '
-    if pkg_resources.get_distribution("sphinx").version.startswith('1.2'):
-        sphinx_opts += '-j ' + str(cpu_count() + 1 ) + ' '
-
     if target.endswith('-nitpick'):
-        sphinx_opts += '-n -w $(branch-output)/build.$(shell date +%Y%m%d%H%M).log '
+        nitpick = True
         builder = target.split('-')[0]
+        fab_args = builder + ',nitpick=True'
     else:
+        nitpick = False
         builder = target
+        fab_args = builder
 
-    sphinx_opts += '$(branch-output)/source'
+    m.job('fab sphinx.build:' + fab_args, block=b)
+    m.msg('[{0}] completed {0} build.'.format(target))
 
-    if target.startswith('epub'):
-        epub_filter = ("sed $(SED_ARGS_REGEX) "
-                       "-e '/^WARNING: unknown mimetype.*ignoring$$/d' "
-                       "-e '/^WARNING: search index.*incomplete.$$/d' ")
 
-        epub_command = 'sphinx-build -b epub -t print {0} $(branch-output)/epub'.format(sphinx_opts)
-
-        m.job('{ %s 2>&1 1>&3 | %s 1>&2; } 3>&1' % (epub_command, epub_filter), block=b)
-    else:
-        m.job('sphinx-build -b {0} -t {1} {2} $(branch-output)/{3}'.format(builder, tag, sphinx_opts, builder), block=b)
-
-    if target.startswith('linkcheck'):
-        m.msg('[' + target + ']: Link check complete at `date`. See "$(branch-output)/linkcheck/output.txt".', block=b)
-    else:
-        m.msg('[' + target + ']: build finished at `date`.', block=b)
-
-    m.target('clean-' + target, block=b)
-    m.job('rm -rf $(branch-output)/doctrees-{0} $(branch-output)/{0}'.format(builder), block=b)
-    m.msg('[clean-{0}]: removed all files supporting the {1} build'.format(target, builder) )
+    if nitpick is False:
+        m.target('clean-' + target, block=b)
+        m.job('rm -rf $(branch-output)/doctrees-{0} $(branch-output)/{0}'.format(builder), block=b)
+        m.msg('[clean-{0}]: removed all files supporting the {1} build'.format(target, builder) )
 
     m.newline(block=b)
